wco_api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
tensorflow>=2.16.0
Pillow>=10.0.0
numpy>=1.26.0
psutil>=5.9.0
python-multipart>=0.0.9
